Rename Hash::Schema to Schema and move parameters from meta to options
Meta can be optionally omitted when dumping to AST, we can't let it happen.
