+++ lib/dry/types/printer.rb
@@ -11,8 +11,8 @@ module Dry
         Constrained => :visit_constrained,
         Constrained::Coercible => :visit_constrained,
         Hash => :visit_hash,
-        Hash::Schema => :visit_schema,
-        Hash::Key => :visit_key,
+        Schema => :visit_schema,
+        Schema::Key => :visit_key,
         Map => :visit_map,
         Array => :visit_array,
         Array::Member => :visit_array_member,
