+++ spec/dry/types/constructor_spec.rb
@@ -148,13 +148,13 @@ RSpec.describe Dry::Types::Constructor do
 
     it "doesn't wrap not composable types" do
       schema = type.schema(age: 'strict.integer').constructor { |input| input.transform_keys(&:to_sym) }
-      expect(schema.key(:age)).to be_a(Dry::Types::Hash::Key)
+      expect(schema.key(:age)).to be_a(Dry::Types::Schema::Key)
     end
 
     it 'chooses the right constructor types' do
       sum = type.schema(age: 'strict.integer').optional
       schema = sum.constructor { |input| input.transform_keys(&:to_sym) if input }
-      expect(schema.right.key(:age)).to be_a(Dry::Types::Hash::Key)
+      expect(schema.right.key(:age)).to be_a(Dry::Types::Schema::Key)
     end
   end
