+++ spec/dry/types/schema/key_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe Dry::Types::Hash::Key do
+RSpec.describe Dry::Types::Schema::Key do
   let(:key) { described_class.new(Dry::Types['strict.integer'], :age) }
   subject { key }
