+++ spec/dry/types/schema_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe Dry::Types::Hash::Schema do
+RSpec.describe Dry::Types::Schema do
   subject(:schema) do
     Dry::Types['strict.hash'].schema(
       name: "coercible.string",
@@ -312,57 +312,6 @@ RSpec.describe Dry::Types::Hash::Schema do
 
       example 'not required' do
         expect(subject.key(:age)).not_to be_required
-      end
-    end
-  end
-
-  describe '#to_s' do
-    context 'inline' do
-      context 'empty schema' do
-        subject(:type) { Dry::Types['hash'].schema({}) }
-
-        it 'returns string representation of the type' do
-          expect(type.to_s).to eql('#<Dry::Types[Schema<keys={}>]>')
-        end
-      end
-
-      context 'strict schema' do
-        subject(:type) { Dry::Types['hash'].schema({}).strict }
-
-        it 'returns string representation of the type' do
-          expect(type.to_s).to eql('#<Dry::Types[Schema<strict keys={}>]>')
-        end
-      end
-
-      context 'key transformation' do
-        let(:key_transformation) { :to_sym.to_proc }
-
-        subject(:type) { Dry::Types['hash'].schema({}).with_key_transform(key_transformation) }
-
-        it 'returns string representation of the type' do
-          expect(type.to_s).
-            to eql("#<Dry::Types[Schema<key_fn=.to_sym keys={}>]>")
-        end
-      end
-
-      context 'type transformation' do
-        let(:type_transformation) { :safe.to_proc }
-
-        subject(:type) { Dry::Types['hash'].with_type_transform(type_transformation).schema({}) }
-
-        it 'returns string representation of the type' do
-          expect(type.to_s).
-            to eql("#<Dry::Types[Schema<type_fn=.safe keys={}>]>")
-        end
-      end
-
-      context 'schema with keys' do
-        subject(:type) { Dry::Types['hash'].schema(name: 'string', age?: 'integer') }
-
-        it 'returns string representation of the type' do
-          expect(type.to_s).
-            to eql("#<Dry::Types[Schema<keys={name: Definition<String> age?: Definition<Integer>}>]>")
-        end
       end
     end
   end
